requirements.txt
asgiref==3.7.2
black==23.3.0
click==8.1.3
Django==4.2.2
flake8==6.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.25.0
opencv-python==4.8.0.74
packaging==23.1
pathspec==0.11.1
Pillow==10.0.0
platformdirs==3.8.0
pycodestyle==2.10.0
pyflakes==3.0.1
sqlparse==0.4.4
